+++ lib/noticed.rb
@@ -6,6 +6,7 @@ module Noticed
   autoload :Base, "noticed/base"
   autoload :Coder, "noticed/coder"
   autoload :Model, "noticed/model"
+  autoload :TextCoder, "noticed/text_coder"
   autoload :Translation, "noticed/translation"
 
   module DeliveryMethods